KAFKA-16223: Replace EasyMock/PowerMock with Mockito for KafkaConfigBackingStoreTest
